+++ tests/test_flac.py
@@ -15,15 +15,18 @@ from audio_metadata import (
 	FLACPicture,
 	FLACSeekPoint,
 	FLACSeekTable,
+	FLACStreamInfo,
+	FLACVorbisComments,
 	FormatError,
 	ID3PictureType,
 )
 from tests.fixtures import (
 	flac_0_duration,
-	flac_application,
+	flac_0_size_block,
 	flac_application_block,
-	flac_cuesheet,
+	flac_application_data,
 	flac_cuesheet_block,
+	flac_cuesheet_data,
 	flac_cuesheet_index_1,
 	flac_cuesheet_index_2,
 	flac_cuesheet_track_1,
@@ -31,135 +34,31 @@ from tests.fixtures import (
 	flac_id3v2,
 	flac_invalid_block,
 	flac_padding_block,
-	flac_picture,
+	flac_padding_data,
 	flac_picture_block,
+	flac_picture_data,
 	flac_reserved_block,
-	flac_seektable,
 	flac_seektable_block,
+	flac_seektable_data,
+	flac_streaminfo_block,
+	flac_streaminfo_data,
 	flac_vorbis_comment_block,
+	flac_vorbis_comment_data,
 	null,
 )
-
-
-@test(
-	"Ignore ID3v2",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_id3v2=flac_id3v2)
-def test_FLAC_load_ID3v2(flac_id3v2):
-	FLAC.parse(flac_id3v2)
-
-
-@test(
-	"Load padding block",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_padding_block=flac_padding_block)
-def test_FLAC_load_padding(flac_padding_block):
-	FLAC.parse(flac_padding_block)
-
-
-@test(
-	"Load application block",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_application_block=flac_application_block)
-def test_FLAC_load_application(flac_application_block):
-	FLAC.parse(flac_application_block)
-
-
-@test(
-	"Load seektable block",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_seektable_block=flac_seektable_block)
-def test_FLAC_load_seektable(flac_seektable_block):
-	FLAC.parse(flac_seektable_block)
-
-
-@test(
-	"Load vorbis comment block",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_vorbis_comment_block=flac_vorbis_comment_block)
-def test_FLAC_load_vorbis_comment(flac_vorbis_comment_block):
-	FLAC.parse(flac_vorbis_comment_block)
-
-
-@test(
-	"Load cuesheet block",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_cuesheet_block=flac_cuesheet_block)
-def test_FLAC_load_cuesheet(flac_cuesheet_block):
-	FLAC.parse(flac_cuesheet_block)
-
-
-@test(
-	"Load picture block",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_picture_block=flac_picture_block)
-def test_FLAC_load_picture(flac_picture_block):
-	FLAC.parse(flac_picture_block)
-
-
-@test(
-	"Reserved block type is a FLACMetadataBlock",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_reserved_block=flac_reserved_block)
-def _(flac_reserved_block):
-	flac = FLAC.parse(flac_reserved_block)
-
-	assert flac._blocks[0] == FLACMetadataBlock(
-		type=10,
-		data=b'',
-	)
-
-
-@test(
-	"Load 0 duration",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_0_duration=flac_0_duration)
-def _(flac_0_duration):
-	FLAC.parse(flac_0_duration)
-
-
-@test(
-	"No FLAC header raises FormatError",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(null=null)
-def _(null):
-	with raises(FormatError) as ctx:
-		FLAC.parse(null)
-	assert str(ctx.raised) == "Valid FLAC header not found."
-
-
-@test(
-	"Invalid block type raises FormatError",
-	tags=['unit', 'flac', 'FLAC'],
-)
-@using(flac_invalid_block=flac_invalid_block)
-def _(flac_invalid_block):
-	with raises(FormatError) as ctx:
-		FLAC.parse(flac_invalid_block)
-	assert str(ctx.raised) == "127 is not a valid FLAC metadata block type."
 
 
 @test(
 	"FLACApplication",
 	tags=['unit', 'flac', 'FLACApplication'],
 )
-@using(flac_application=flac_application)
-def _(flac_application):
+@using(flac_application_data=flac_application_data)
+def _(flac_application_data):
 	application_init = FLACApplication(
 		id='aiff',
 		data=b'FORM\x02\xe0\x9b\x08AIFF'
 	)
-	application_load = FLACApplication.parse(flac_application)
+	application_load = FLACApplication.parse(flac_application_data)
 
 	assert application_init == application_load
 	assert application_init.id == application_load.id == 'aiff'
@@ -263,8 +162,8 @@ def _(flac_cuesheet_track_1, flac_cuesheet_track_2):
 	"FLACCueSheet",
 	tags=['unit', 'flac', 'FLACCueSheet'],
 )
-@using(flac_cuesheet=flac_cuesheet)
-def _(flac_cuesheet):
+@using(flac_cuesheet_data=flac_cuesheet_data)
+def _(flac_cuesheet_data):
 	cuesheet_init = FLACCueSheet(
 		[
 			FLACCueSheetTrack(
@@ -323,7 +222,7 @@ def _(flac_cuesheet):
 		lead_in_samples=88200,
 		compact_disc=True,
 	)
-	cuesheet_load = FLACCueSheet.parse(flac_cuesheet)
+	cuesheet_load = FLACCueSheet.parse(flac_cuesheet_data)
 
 	assert cuesheet_init == cuesheet_load
 	assert repr(cuesheet_init) == repr(cuesheet_load) == '<FLACCueSheet (4 tracks)>'
@@ -348,9 +247,10 @@ def _():
 	"FLACPadding",
 	tags=['unit', 'flac', 'FLACPadding'],
 )
-def _():
+@using(flac_padding_data=flac_padding_data)
+def _(flac_padding_data):
 	padding_init = FLACPadding(size=10)
-	padding_load = FLACPadding.parse(b'\x00' * 10)
+	padding_load = FLACPadding.parse(flac_padding_data)
 
 	assert padding_init == padding_load
 	assert repr(padding_init) == repr(padding_load) == '<FLACPadding (10 bytes)>'
@@ -360,9 +260,9 @@ def _():
 	"FLACPicture",
 	tags=['unit', 'flac', 'FLACPicture'],
 )
-@using(flac_picture=flac_picture)
-def _(flac_picture):
-	vorbis_picture_init = FLACPicture(
+@using(flac_picture_data=flac_picture_data)
+def _(flac_picture_data):
+	flac_picture_init = FLACPicture(
 		type=ID3PictureType.COVER_FRONT,
 		mime_type='image/png',
 		description='',
@@ -370,19 +270,19 @@ def _(flac_picture):
 		height=16,
 		bit_depth=32,
 		colors=0,
-		data=flac_picture[41:],
+		data=flac_picture_data[41:],
 	)
-	vorbis_picture_load = FLACPicture.parse(flac_picture)
+	flac_picture_load = FLACPicture.parse(flac_picture_data)
 
-	assert vorbis_picture_init == vorbis_picture_load
+	assert flac_picture_init == flac_picture_load
 
 
 @test(
 	"FLACSeekTable",
 	tags=['unit', 'flac', 'FLACSeekTable'],
 )
-@using(flac_seektable=flac_seektable)
-def _(flac_seektable):
+@using(flac_seektable_data=flac_seektable_data)
+def _(flac_seektable_data):
 	seekpoints = [
 		FLACSeekPoint(
 			first_sample=first_sample,
@@ -398,7 +298,393 @@ def _(flac_seektable):
 	]
 
 	seektable_init = FLACSeekTable(seekpoints)
-	seektable_load = FLACSeekTable.parse(flac_seektable)
+	seektable_load = FLACSeekTable.parse(flac_seektable_data)
 
 	assert seektable_init == seektable_load
 	assert seektable_init.data == seektable_load.data == seekpoints
+
+
+@test(
+	"FLACStreamInfo",
+	tags=['unit', 'flac', 'FLACStreamInfo'],
+)
+@using(flac_streaminfo_data=flac_streaminfo_data)
+def _(flac_streaminfo_data):
+	flac_streaminfo_init = FLACStreamInfo(
+		start=None,
+		size=None,
+		min_block_size=4096,
+		max_block_size=4096,
+		min_frame_size=14,
+		max_frame_size=16,
+		bit_depth=16,
+		bitrate=None,
+		channels=2,
+		duration=5,
+		md5='9b1be87c6b579fde2341515f4d82c008',
+		sample_rate=44100,
+	)
+	flac_streaminfo_load = FLACStreamInfo.parse(flac_streaminfo_data)
+
+	assert flac_streaminfo_init == flac_streaminfo_load
+
+
+@test(
+	"FLACVorbisComments",
+	tags=['unit', 'flac', 'FLACVorbisComments'],
+)
+@using(flac_vorbis_comment_data=flac_vorbis_comment_data)
+def _(flac_vorbis_comment_data):
+	vorbis_comments = FLACVorbisComments.parse(flac_vorbis_comment_data)
+
+	assert vorbis_comments == FLACVorbisComments(
+		_vendor='reference libFLAC 1.3.2 20170101',
+		album=['test-album'],
+		artist=['test-artist'],
+		comment=['test-comment'],
+		date=['2000'],
+		discnumber=['1'],
+		disctotal=['99'],
+		genre=['test-genre'],
+		title=['test-title'],
+		tracknumber=['1'],
+		tracktotal=['99'],
+	)
+
+
+@test(
+	"Parse FLAC application block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_application_block=flac_application_block)
+def _(flac_application_block):
+	application_block, _ = FLAC._parse_metadata_block(flac_application_block)
+
+	assert application_block == FLACApplication(
+		id='aiff',
+		data=b'FORM\x02\xe0\x9b\x08AIFF'
+	)
+
+
+@test(
+	"Parse FLAC cuesheet block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_cuesheet_block=flac_cuesheet_block)
+def _(flac_cuesheet_block):
+	cuesheet_block, _ = FLAC._parse_metadata_block(flac_cuesheet_block)
+
+	assert cuesheet_block == FLACCueSheet(
+		[
+			FLACCueSheetTrack(
+				track_number=1,
+				offset=0,
+				isrc='123456789012',
+				type=0,
+				pre_emphasis=False,
+				indexes=[
+					FLACCueSheetIndex(
+						number=1,
+						offset=0,
+					)
+				]
+			),
+			FLACCueSheetTrack(
+				track_number=2,
+				offset=44100,
+				isrc='',
+				type=1,
+				pre_emphasis=True,
+				indexes=[
+					FLACCueSheetIndex(
+						number=1,
+						offset=0,
+					),
+					FLACCueSheetIndex(
+						number=2,
+						offset=588,
+					),
+				],
+			),
+			FLACCueSheetTrack(
+				track_number=3,
+				offset=88200,
+				isrc='',
+				type=0,
+				pre_emphasis=False,
+				indexes=[
+					FLACCueSheetIndex(
+						number=1,
+						offset=0,
+					)
+				],
+			),
+			FLACCueSheetTrack(
+				track_number=170,
+				offset=162496,
+				isrc='',
+				type=0,
+				pre_emphasis=False,
+				indexes=[]
+			)
+		],
+		catalog_number='1234567890123',
+		lead_in_samples=88200,
+		compact_disc=True,
+	)
+
+
+@test(
+	"Parse FLAC padding block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_padding_block=flac_padding_block)
+def _(flac_padding_block):
+	padding_block, _ = FLAC._parse_metadata_block(flac_padding_block)
+
+	assert padding_block == FLACPadding(size=10)
+
+
+@test(
+	"Parse FLAC picture block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_picture_block=flac_picture_block)
+def _(flac_picture_block):
+	picture_block, _ = FLAC._parse_metadata_block(flac_picture_block)
+
+	assert picture_block == FLACPicture(
+		type=ID3PictureType.COVER_FRONT,
+		mime_type='image/png',
+		description='',
+		width=16,
+		height=16,
+		bit_depth=32,
+		colors=0,
+		data=flac_picture_block[45:],
+	)
+
+
+@test(
+	"Parse FLAC seektable block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_seektable_block=flac_seektable_block)
+def _(flac_seektable_block):
+	seektable_block, _ = FLAC._parse_metadata_block(flac_seektable_block)
+
+	assert seektable_block == FLACSeekTable(
+		[
+			FLACSeekPoint(
+				first_sample=first_sample,
+				offset=offset,
+				num_samples=num_samples)
+			for first_sample, offset, num_samples in [
+				(0, 0, 4096),
+				(40960, 140, 4096),
+				(86016, 294, 4096),
+				(131072, 448, 4096),
+				(176128, 602, 4096)
+			]
+		]
+	)
+
+
+@test(
+	"Parse FLAC streaminfo block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_streaminfo_block=flac_streaminfo_block)
+def _(flac_streaminfo_block):
+	streaminfo_block, _ = FLAC._parse_metadata_block(flac_streaminfo_block)
+
+	assert streaminfo_block == FLACStreamInfo(
+		start=None,
+		size=None,
+		min_block_size=4096,
+		max_block_size=4096,
+		min_frame_size=14,
+		max_frame_size=16,
+		bit_depth=16,
+		bitrate=None,
+		channels=2,
+		duration=5,
+		md5='9b1be87c6b579fde2341515f4d82c008',
+		sample_rate=44100,
+	)
+
+
+@test(
+	"Parse FLAC vorbis comment block",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_vorbis_comment_block=flac_vorbis_comment_block)
+def _(flac_vorbis_comment_block):
+	vorbis_comment_block, _ = FLAC._parse_metadata_block(flac_vorbis_comment_block)
+
+	assert vorbis_comment_block == FLACVorbisComments(
+		_vendor='reference libFLAC 1.3.2 20170101',
+		album=['test-album'],
+		artist=['test-artist'],
+		comment=['test-comment'],
+		date=['2000'],
+		discnumber=['1'],
+		disctotal=['99'],
+		genre=['test-genre'],
+		title=['test-title'],
+		tracknumber=['1'],
+		tracktotal=['99'],
+	)
+
+
+@test(
+	"0-sized FLAC block type raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_0_size_block=flac_0_size_block)
+def _(flac_0_size_block):
+	with raises(FormatError) as ctx:
+		FLAC._parse_metadata_block(flac_0_size_block)
+	assert str(ctx.raised) == "FLAC metadata block size must be greater than 0."
+
+
+@test(
+	"Invalid FLAC block type raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_invalid_block=flac_invalid_block)
+def _(flac_invalid_block):
+	with raises(FormatError) as ctx:
+		FLAC._parse_metadata_block(flac_invalid_block)
+	assert str(ctx.raised) == "127 is not a valid FLAC metadata block type."
+
+
+@test(
+	"Flac reserved block type is a FLACMetadataBlock",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_reserved_block=flac_reserved_block)
+def _(flac_reserved_block):
+	reserved_block, _ = FLAC._parse_metadata_block(flac_reserved_block)
+
+	assert reserved_block == FLACMetadataBlock(
+		type=10,
+		data=b'\x00\x00\x00\x00',
+	)
+
+
+@test(
+	"Ignore ID3v2 in FLAC",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_id3v2=flac_id3v2)
+def _(flac_id3v2):
+	FLAC.parse(flac_id3v2)
+
+
+@test(
+	"0 duration FLAC",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_0_duration=flac_0_duration)
+def _(flac_0_duration):
+	FLAC.parse(flac_0_duration)
+
+
+@test(
+	"Reserved FLAC metadata block type",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(
+	flac_reserved_block=flac_reserved_block,
+	flac_streaminfo_block=flac_streaminfo_block,
+)
+def _(
+	flac_reserved_block,
+	flac_streaminfo_block,
+):
+	FLAC.parse(b'fLaC' + flac_streaminfo_block + b'\x8a' + flac_reserved_block[1:])
+
+
+@test(
+	"No FLAC header raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(null=null)
+def _(null):
+	with raises(FormatError) as ctx:
+		FLAC.parse(null)
+	assert str(ctx.raised) == "Valid FLAC header not found."
+
+
+@test(
+	"FLAC streaminfo block not first raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_padding_block=flac_padding_block)
+def _(flac_padding_block):
+	with raises(FormatError) as ctx:
+		FLAC.parse(b'fLaC' + flac_padding_block)
+	assert str(ctx.raised) == "FLAC streaminfo block must be first."
+
+
+@test(
+	"Multiple FLAC cuesheet blocks raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(
+	flac_cuesheet_block=flac_cuesheet_block,
+	flac_streaminfo_block=flac_streaminfo_block,
+)
+def _(
+	flac_cuesheet_block,
+	flac_streaminfo_block,
+):
+	with raises(FormatError) as ctx:
+		FLAC.parse(b'fLaC' + flac_streaminfo_block + flac_cuesheet_block + flac_cuesheet_block)
+	assert str(ctx.raised) == "Multiple FLAC cuesheet blocks found."
+
+
+@test(
+	"Multiple FLAC seektable blocks raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(
+	flac_seektable_block=flac_seektable_block,
+	flac_streaminfo_block=flac_streaminfo_block,
+)
+def _(
+	flac_seektable_block,
+	flac_streaminfo_block,
+):
+	with raises(FormatError) as ctx:
+		FLAC.parse(b'fLaC' + flac_streaminfo_block + flac_seektable_block + flac_seektable_block)
+	assert str(ctx.raised) == "Multiple FLAC seektable blocks found."
+
+
+@test(
+	"Multiple FLAC streaminfo blocks raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(flac_streaminfo_block=flac_streaminfo_block)
+def _(flac_streaminfo_block):
+	with raises(FormatError) as ctx:
+		FLAC.parse(b'fLaC' + flac_streaminfo_block + flac_streaminfo_block)
+	assert str(ctx.raised) == "Multiple FLAC streaminfo blocks found."
+
+
+@test(
+	"Multiple FLAC Vorbis comment blocks raises FormatError",
+	tags=['unit', 'flac', 'FLAC'],
+)
+@using(
+	flac_vorbis_comment_block=flac_vorbis_comment_block,
+	flac_streaminfo_block=flac_streaminfo_block,
+)
+def _(
+	flac_vorbis_comment_block,
+	flac_streaminfo_block,
+):
+	with raises(FormatError) as ctx:
+		FLAC.parse(b'fLaC' + flac_streaminfo_block + flac_vorbis_comment_block + flac_vorbis_comment_block)
+	assert str(ctx.raised) == "Multiple FLAC Vorbis comment blocks found."